// src/app/interfaces/interfaces.ts
export interface Usuario {
  _id?:      string;
  nombre?:   string;
  avatar?:   string;
  email?:    string;
  password?:    string;
  profile?: string;
}

export interface LoginForm{
  email:    string;
  password:    string;
}


// Generated by https://quicktype.io

export interface LoginResponse {
  user:  User;
  token: string;
}

export interface User {
  id:                  string;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  escuela_id:          null;
  name:                string;
  email:               string;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  v_profile_img:       string;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  email_verified_at:   null;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  d_expire_password:   null;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  i_retry:             number;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  i_lockout:           number;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  b_unread_chat:       number;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  b_unread_comunicado: number;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  v_profile:           string;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  v_status:            string;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  created_at:          string;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  updated_at:          string;
}